+++ packages/react-tinacms-github/src/github-editing-context/TinacmsGithubProvider.tsx
@@ -18,10 +18,9 @@ limitations under the License.
 
 import React, { useState, useEffect } from 'react'
 import { useCMS } from 'tinacms'
-import GithubErrorModal from '../github-error/GithubErrorModal'
+import GithubErrorModal, { GithubError } from '../github-error/GithubErrorModal'
 import { CreateForkModal, GithubAuthenticationModal } from './GithubAuthModal'
 import { GithubEditingContext } from './GithubEditingContext'
-import { useGithubEditing } from './useGithubEditing'
 import { GithubClient } from '../github-client'
 
 interface ProviderProps {
@@ -41,7 +40,7 @@ export const TinacmsGithubProvider = ({
   exitEditMode,
   error: previewError,
 }: ProviderProps) => {
-  const [error, setError] = useState<any>(null)
+  const [error, setError] = useState<GithubError>(previewError)
   const cms = useCMS()
   const github: GithubClient = cms.api.github
   const [activeModal, setActiveModal] = useState<ModalNames>(null)
@@ -60,11 +59,18 @@ export const TinacmsGithubProvider = ({
 
   const onAuthSuccess = async () => {
     if (await github.isAuthorized()) {
+      github.setWorkingRepoFullName(github.baseRepoFullName)
       enterEditMode()
     } else {
       setActiveModal('createFork')
     }
   }
+
+  useEffect(() => {
+    return cms.events.subscribe('github:error', ({ error }) => {
+      setError(error)
+    })
+  })
 
   return (
     <GithubEditingContext.Provider
@@ -72,7 +78,6 @@ export const TinacmsGithubProvider = ({
         editMode,
         enterEditMode: beginAuth,
         exitEditMode,
-        setError,
       }}
     >
       {error && <GithubErrorModal error={error} />}
@@ -85,33 +90,7 @@ export const TinacmsGithubProvider = ({
       {!error && activeModal === 'createFork' && (
         <CreateForkModal close={onClose} onForkCreated={enterEditMode} />
       )}
-      <PreviewErrorBoundary previewError={previewError}>
-        {children}
-      </PreviewErrorBoundary>
+      {!previewError && children}
     </GithubEditingContext.Provider>
   )
-}
-
-interface Props {
-  previewError: any
-  children: any
-}
-function PreviewErrorBoundary(props: Props) {
-  const github = useGithubEditing()
-
-  useEffect(() => {
-    ;(async () => {
-      if (props.previewError) {
-        github.setError(props.previewError)
-      }
-    })()
-  }, [props.previewError])
-
-  if (props.previewError) {
-    return null
-  }
-
-  // don't show content with initial content error
-  // because the data is likely missing
-  return props.children
 }