+++ packages/gatsby-tinacms-json/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.21-alpha.3](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.21-alpha.2...gatsby-tinacms-json@0.8.21-alpha.3) (2020-05-08)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.21-alpha.2](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.21-alpha.1...gatsby-tinacms-json@0.8.21-alpha.2) (2020-05-08)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.21-alpha.1](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.21-alpha.0...gatsby-tinacms-json@0.8.21-alpha.1) (2020-05-08)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
+## [0.8.21-alpha.0](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.20...gatsby-tinacms-json@0.8.21-alpha.0) (2020-05-06)
+
+**Note:** Version bump only for package gatsby-tinacms-json
+
+
+
+
+
 ## [0.8.20](https://github.com/tinacms/tinacms/compare/gatsby-tinacms-json@0.8.20-alpha.0...gatsby-tinacms-json@0.8.20) (2020-05-04)