+++ packages/next-tinacms-markdown/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.14-alpha.3](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.14-alpha.2...next-tinacms-markdown@0.2.14-alpha.3) (2020-05-08)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
+## [0.2.14-alpha.2](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.14-alpha.1...next-tinacms-markdown@0.2.14-alpha.2) (2020-05-08)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
+## [0.2.14-alpha.1](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.14-alpha.0...next-tinacms-markdown@0.2.14-alpha.1) (2020-05-08)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
+## [0.2.14-alpha.0](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.13...next-tinacms-markdown@0.2.14-alpha.0) (2020-05-06)
+
+**Note:** Version bump only for package next-tinacms-markdown
+
+
+
+
+
 ## [0.2.13](https://github.com/tinacms/tinacms/compare/next-tinacms-markdown@0.2.13-alpha.0...next-tinacms-markdown@0.2.13) (2020-05-04)
 
 **Note:** Version bump only for package next-tinacms-markdown