+++ packages/tinacms/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.19.2-alpha.3](https://github.com/tinacms/tinacms/compare/tinacms@0.19.2-alpha.2...tinacms@0.19.2-alpha.3) (2020-05-08)
+
+**Note:** Version bump only for package tinacms
+
+
+
+
+
+## [0.19.2-alpha.2](https://github.com/tinacms/tinacms/compare/tinacms@0.19.2-alpha.1...tinacms@0.19.2-alpha.2) (2020-05-08)
+
+**Note:** Version bump only for package tinacms
+
+
+
+
+
+## [0.19.2-alpha.1](https://github.com/tinacms/tinacms/compare/tinacms@0.19.2-alpha.0...tinacms@0.19.2-alpha.1) (2020-05-08)
+
+**Note:** Version bump only for package tinacms
+
+
+
+
+
+## [0.19.2-alpha.0](https://github.com/tinacms/tinacms/compare/tinacms@0.19.1...tinacms@0.19.2-alpha.0) (2020-05-06)
+
+**Note:** Version bump only for package tinacms
+
+
+
+
+
 ## [0.19.1](https://github.com/tinacms/tinacms/compare/tinacms@0.19.1-alpha.0...tinacms@0.19.1) (2020-05-04)
 
 **Note:** Version bump only for package tinacms